requirements.txt
boto3==1.4.4
botocore==1.5.85
PyYAML==3.12
